requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.8.0
Pillow>=9.0.0
matplotlib>=3.5.0
scikit-image>=0.19.0
tqdm>=4.64.0
tensorboard>=2.10.0
albumentations>=1.3.0
